Revert "moved AbstractWorkflowBasedRunConfiguration to de.uka.ipd.sdq.workflow.launchconfig.core"
This reverts commit f62e4da1eba267407eb8c33f28e44563597db0b9.
