+++ bundles/de.uka.ipd.sdq.workflow.launchconfig/META-INF/MANIFEST.MF
@@ -16,8 +16,7 @@ Require-Bundle: org.eclipse.ui,
  org.apache.log4j;bundle-version="1.2.13";visibility:=reexport,
  org.apache.commons.logging;bundle-version="1.1.1",
  de.uka.ipd.sdq.workflow.logging.console;bundle-version="1.0.0";visibility:=reexport,
- de.uka.ipd.sdq.workflow.ui;bundle-version="1.0.0";visibility:=reexport,
- de.uka.ipd.sdq.workflow.launchconfig.core;bundle-version="5.2.0"
+ de.uka.ipd.sdq.workflow.ui;bundle-version="1.0.0";visibility:=reexport
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: sdq.ipd.uka.de
 Export-Package: de.uka.ipd.sdq.workflow.launchconfig,
