+++ bundles/de.uka.ipd.sdq.workflow.mdsd/META-INF/MANIFEST.MF
@@ -31,7 +31,6 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.4.0",
  org.eclipse.m2m.qvt.oml.project;bundle-version="3.1.0",
  org.eclipse.core.resources;bundle-version="3.5.0",
  org.eclipse.emf.ecore.xmi;bundle-version="2.5.0",
- de.uka.ipd.sdq.workflow.launchconfig.core,
  de.uka.ipd.sdq.workflow.launchconfig,
  org.eclipse.emf.mwe.utils,
  com.google.inject;bundle-version="2.0.0",
