+++ bundles/de.uka.ipd.sdq.workflow.mdsd/src/de/uka/ipd/sdq/workflow/mdsd/validation/ShowValidationErrorsJob.java
@@ -14,7 +14,7 @@ import de.uka.ipd.sdq.workflow.jobs.CleanupFailedException;
 import de.uka.ipd.sdq.workflow.jobs.IJob;
 import de.uka.ipd.sdq.workflow.jobs.JobFailedException;
 import de.uka.ipd.sdq.workflow.jobs.UserCanceledException;
-import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowBasedRunConfiguration;
+import de.uka.ipd.sdq.workflow.launchconfig.AbstractWorkflowBasedRunConfiguration;
 
 /**
  * The Class ShowValidationErrorsJob.
